backEnd_Python/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-cors-headers==4.7.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
setuptools==75.8.0
simplejson==3.19.3
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
